Moved `follow` handler to dedicated class
no ref

Moved `follow` handler to dedicated class to follow the same pattern as
the `create` & `delete` handlers in preparation for further work re:
rejecting follow requests for blocked accounts
